Move request line
